feat: screening progress stat

+++ myopia-business/business-api/screening-app-api/src/main/java/com/wupol/myopia/business/api/screening/app/domain/vo/ClassScreeningProgress.java
@@ -13,13 +13,13 @@ import java.util.List;
 @Data
 public class ClassScreeningProgress {
     /** 计划筛查人数 */
-    private Long planCount;
+    private Integer planCount;
     /** 实际筛查人数 */
-    private Long screeningCount;
+    private Integer screeningCount;
     /** 有异常筛查人数，仅统计有初步结果的：眼位、视力检查、电脑验光、小瞳验光 */
-    private Long abnormalCount;
+    private Integer abnormalCount;
     /** 筛查未完成学生数 */
-    private Long unfinishedCount;
+    private Integer unfinishedCount;
     /** 学龄段：5-幼儿园、0-小学、1-初中、2-高中、3-职业高中 */
     private Integer schoolAge;
     /** 当前班级的学生筛查进度情况 */
